+++ __tests__/social.test.js
@@ -1,10 +1,10 @@
 const supertest = require("supertest");
-const server = require("../api/server.js");
+const server = require("../index");
 
 describe("social routes", () => {
     it (" will not let you access social route without a token", async () => {
            await supertest(server)
                 .get("/api/social")
-                .expect(401) 
+                .expect(401)
     })
 })
